+++ delta-lake/delta-42x/src/main/scala/com/nvidia/spark/rapids/delta/delta42x/OptimizeTableCommandMeta.scala
@@ -1,0 +1,78 @@
+/*
+ * Copyright (c) 2026, NVIDIA CORPORATION.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.nvidia.spark.rapids.delta.delta42x
+
+import com.nvidia.spark.rapids.{DataFromReplacementRule, RapidsConf, RapidsMeta}
+import com.nvidia.spark.rapids.delta.RapidsDeltaUtils
+import com.nvidia.spark.rapids.delta.common.OptimizeTableCommandMetaBase
+
+import org.apache.spark.sql.SparkSession
+import org.apache.spark.sql.delta.DeltaLog
+import org.apache.spark.sql.delta.commands.{DeletionVectorUtils, DeltaCommand,
+  OptimizeTableCommand}
+import org.apache.spark.sql.delta.rapids.GpuOptimizeTableCommand
+import org.apache.spark.sql.delta.sources.DeltaSQLConf
+import org.apache.spark.sql.execution.command.RunnableCommand
+
+class OptimizeTableCommandMeta(
+    cmd: OptimizeTableCommand,
+    conf: RapidsConf,
+    parent: Option[RapidsMeta[_, _, _]],
+    rule: DataFromReplacementRule)
+  extends OptimizeTableCommandMetaBase(cmd, conf, parent, rule) {
+
+  private object DeltaCmdProxy extends DeltaCommand
+
+  override protected def getDeltaLogForOptimize(): DeltaLog = {
+    DeltaCmdProxy.getDeltaTable(cmd.child, "OPTIMIZE").deltaLog
+  }
+
+  override def tagSelfForGpu(): Unit = {
+    if (!conf.isDeltaWriteEnabled) {
+      willNotWorkOnGpu("Delta Lake output acceleration has been disabled. To enable set " +
+        s"${RapidsConf.ENABLE_DELTA_WRITE} to true")
+    }
+
+    val deltaLog = getDeltaLogForOptimize()
+    val snapshot = deltaLog.unsafeVolatileSnapshot
+
+    if (DeletionVectorUtils.deletionVectorsWritable(snapshot) &&
+        cmd.conf.getConf(DeltaSQLConf.DELETE_USE_PERSISTENT_DELETION_VECTORS)) {
+      willNotWorkOnGpu("Deletion vectors are not supported on GPU")
+    }
+
+    if (cmd.zOrderBy.nonEmpty) {
+      willNotWorkOnGpu("Z-Order optimize is not supported on GPU")
+    }
+
+    RapidsDeltaUtils.tagForDeltaWrite(
+      this,
+      snapshot.schema,
+      Some(deltaLog),
+      Map.empty,
+      SparkSession.active)
+
+    if (snapshot.isCatalogOwned) {
+      willNotWorkOnGpu("Delta 4.2 requires catalog-managed OPTIMIZE to run on CPU")
+    }
+  }
+
+  override def convertToGpu(): RunnableCommand = {
+    GpuOptimizeTableCommand(cmd.child, cmd.userPartitionPredicates, cmd.optimizeContext)(
+      cmd.zOrderBy)
+  }
+}
